backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Konstantin Pershin <kostya.pershin.18@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
fastapi-jsonrpc = "^3.1.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
SQLAlchemy = "^2.0.30"
bcrypt = "^4.1.3"
PyJWT = "^2.8.0"
celery = "^5.4.0"
redis = "^5.0.4"
pytz = "^2024.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
